change type double to uint64_t

+++ src/evict/evict_lru.c
@@ -738,7 +738,7 @@ __evict_update_work(WT_SESSION_IMPL *session)
      * If application threads are blocked by data in cache, track the fill ratio.
      *
      */
-    double cache_fill_ratio = bytes_inuse / bytes_max;
+    uint64_t cache_fill_ratio = bytes_inuse / bytes_max;
     bool evict_is_hard = LF_ISSET(WT_EVICT_CACHE_HARD);
     if (evict_is_hard) {
         if (cache_fill_ratio < 0.25)
